refactor(optimiser): move all optimiser code into single serverless endpoint
