+++ website/api/optimiser/optimise/modules/modules.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/nusmodifications/nusmods/optimiser/lib/models"
+	"github.com/nusmodifications/nusmods/website/api/optimiser/optimise/models"
 )
 
 var E_Venues = map[string]bool{
